+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2021-07-30
+
+### Changed
+
+- added universal project
+- removed webchat, token storage and many more things
 ## [3.2.2] - 10.12.2020
 ### Fixed
 - removed anonymize from the bot.js